saved model from training_deploy.py
