+++ DL/training_sequence.py
@@ -1,3 +1,4 @@
+#val : 0.66
 # To add a new cell, type '# %%'
 # To add a new markdown cell, type '# %% [markdown]'
 # %%
@@ -34,18 +35,18 @@ elif user == 'aws':
     path_labels_csv = '/home/ubuntu/Data/labels_framewise_csv.csv'
     path_labels_list = '/home/ubuntu/Data/labels_framewise_list.pkl'
     path_frames = '/home/ubuntu/Data/Frames/'
-    checkpoint_path = "/home/ubuntu/checkpoints/training_lstm/cp.ckpt"
+    checkpoint_path = "/home/ubuntu/checkpoints/training_sequence/cp.ckpt"
 
 
 # %%
-#Perform train-test-validation split(66-22-16)
+#Perform train-test-validation split(68-20-16)
 
 x = np.arange(1, 105)
-np.random.seed(5)
+np.random.seed(42)
 np.random.shuffle(x)
 videos_validation = x[:16]
-videos_test = x[16: 16+22]
-videos_train = x[16+22: ]
+videos_test = x[16: 16+20]
+videos_train = x[16+20: ]
 
 print(videos_train, len(videos_train))
 print(videos_test, len(videos_test))
@@ -58,7 +59,7 @@ filenames_validation = []
 labels_validation = []
 filenames_test = []
 labels_test = []
-w = 5
+w = 3
 
 for vid in videos_train:
     folder = path_frames + "video{}/".format(vid)
@@ -117,28 +118,6 @@ print(filenames_train.shape, filenames_validation.shape, filenames_test.shape)
 print(labels_train.shape, labels_validation.shape, labels_test.shape)
 
 
-#%%
-
-ind0 = np.where(labels_train==0)[0]
-ind1 = np.where(labels_train==1)[0]
-random.shuffle(ind0)
-random.shuffle(ind1)
-
-if (ind0.shape[0]/ind1.shape[0] > 1.4):
-    print('reducing the number of unsafe frames in dataframe\n\n')
-    len_ind0 = int(ind1.shape[0]*1.4)
-    ind0 = ind0[:len_ind0]
-
-    indices_required = np.concatenate((ind0, ind1))
-
-filenames_train_reduced = filenames_train[indices_required]
-labels_train_reduced = labels_train[indices_required]
-
-print(filenames_train_reduced.shape, labels_train_reduced.shape)
-
-print(ind0.shape, ind1.shape)
-
-
 # %%
 # Generators
 def parse_function(filename, label):
@@ -146,9 +125,9 @@ def parse_function(filename, label):
     image = tf.io.read_file(filename)
     image = tf.image.decode_jpeg(image)
     image = tf.image.convert_image_dtype(image, tf.float32)
-    image = tf.image.resize(image, [90, 160], method=tf.image.ResizeMethod.AREA, 
+    image = tf.image.resize(image, [135, 240], method=tf.image.ResizeMethod.AREA, 
                             preserve_aspect_ratio=True)
-    #image = tf.image.random_crop(image, size=[270,270,3])
+    #image = image / 255.0
     return image, label
 
 
@@ -164,7 +143,7 @@ dataset_train = tf.data.Dataset.from_tensor_slices((filenames_train,labels_train
 dataset_train = dataset_train.map(parse_function, num_parallel_calls=4)
 dataset_train = dataset_train.map(train_preprocess, num_parallel_calls=4)
 dataset_train = dataset_train.window(w)
-dataset_train = dataset_train.shuffle(len(filenames_train))
+dataset_train = dataset_train.shuffle(int(len(filenames_train)/3))
 dataset_train = dataset_train.flat_map(lambda a,b:tf.data.Dataset.zip((a,b)).batch(8))
 dataset_train = dataset_train.map(lambda a,b : (a,b[-1]))
 dataset_train = dataset_train.batch(8)
@@ -174,7 +153,7 @@ dataset_test = tf.data.Dataset.from_tensor_slices((filenames_test,labels_test))
 dataset_test = dataset_test.map(parse_function, num_parallel_calls=4)
 #dataset_test = dataset_test.map(train_preprocess, num_parallel_calls=4)
 dataset_test = dataset_test.window(w)
-dataset_test = dataset_test.shuffle(len(filenames_test))
+dataset_test = dataset_test.shuffle(int(len(filenames_test)/3))
 dataset_test = dataset_test.flat_map(lambda a,b:tf.data.Dataset.zip((a,b)).batch(8))
 dataset_test = dataset_test.map(lambda a,b : (a,b[-1]))
 dataset_test = dataset_test.batch(8)
@@ -184,7 +163,7 @@ dataset_val = tf.data.Dataset.from_tensor_slices((filenames_validation,labels_va
 dataset_val = dataset_val.map(parse_function, num_parallel_calls=4)
 #dataset_val = dataset_val.map(train_preprocess, num_parallel_calls=4)
 dataset_val = dataset_val.window(w)
-dataset_val = dataset_val.shuffle(len(filenames_validation))
+dataset_val = dataset_val.shuffle(int(len(filenames_validation)/3))
 dataset_val = dataset_val.flat_map(lambda a,b:tf.data.Dataset.zip((a,b)).batch(8))
 dataset_val = dataset_val.map(lambda a,b : (a,b[-1]))
 dataset_val = dataset_val.batch(8)
@@ -193,86 +172,72 @@ dataset_val = dataset_val.prefetch(1)
 # %%
 tf.keras.backend.set_image_data_format('channels_last')
 
-def create_model():
-
-    inputs = tf.keras.layers.Input([90, 160, 3])
-    #x = tf.keras.layers.experimental.preprocessing.Normalization(inputs)
-    x = tf.keras.layers.BatchNormalization()(inputs)
-    x = tf.keras.layers.Conv2D(32, (7,7), padding='same', activation='relu')(inputs)
-    x = tf.keras.layers.Conv2D(32, (7,7), padding='same', activation='relu')(x)
-    x = tf.keras.layers.BatchNormalization()(x)
-    x = tf.keras.layers.Dropout(0.3)(x)
-    x = tf.keras.layers.MaxPool2D(pool_size=(2,2))(x)
-    x = tf.keras.layers.Conv2D(64, (5,5), padding='same', activation='relu')(x)
-    x = tf.keras.layers.Conv2D(64, (5,5), padding='same', activation='relu')(x)
-    x = tf.keras.layers.BatchNormalization()(x)
-    x = tf.keras.layers.Dropout(0.3)(x)
-    x = tf.keras.layers.MaxPool2D(pool_size=(2,2))(x)
-    x = tf.keras.layers.Conv2D(64, (5,5), padding='same', activation='relu')(x)
-    x = tf.keras.layers.Conv2D(128  , (5,5), padding='same', activation='relu')(x)
-    x = tf.keras.layers.BatchNormalization()(x)
-    x = tf.keras.layers.Dropout(0.3)(x)
-    x = tf.keras.layers.MaxPool2D(pool_size=(2,2))(x)
-    x = tf.keras.layers.Conv2D(128, (3,3), padding='same', activation='relu')(x)
-    x = tf.keras.layers.BatchNormalization()(x)
-    x = tf.keras.layers.GlobalAveragePooling2D()(x)
-    x = tf.keras.layers.Dense(64, activation='relu', kernel_regularizer=tf.keras.regularizers.l2(1e-3))(x)
-    x = tf.keras.layers.Dropout(0.4)(x)
-    x = tf.keras.layers.Dense(32, activation='relu', kernel_regularizer=tf.keras.regularizers.l2(1e-3))(x)
-    x = tf.keras.layers.Dropout(0.4)(x)
-    outputs = tf.keras.layers.Dense(1, activation='sigmoid')(x)
-    model = tf.keras.Model(inputs, outputs)
+def build_convnet(shape=(135, 240, 3)):
+    momentum = 0.9
+    model = tf.keras.Sequential()
+    model.add(tf.keras.layers.Conv2D(8, (3,3), input_shape=shape, padding='same', activation='relu'))
+    model.add(tf.keras.layers.Conv2D(8, (3,3), padding='same', activation='relu'))
+    model.add(tf.keras.layers.BatchNormalization(momentum=momentum))
     
+    model.add(tf.keras.layers.MaxPool2D())
     
-
+    # model.add(tf.keras.layers.Conv2D(128, (3,3), padding='same', activation='relu'))
+    # model.add(tf.keras.layers.Conv2D(128, (3,3), padding='same', activation='relu'))
+    # model.add(tf.keras.layers.BatchNormalization(momentum=momentum))
+    
+    # model.add(tf.keras.layers.MaxPool2D())
+    
+    model.add(tf.keras.layers.Conv2D(16, (3,3), padding='same', activation='relu'))
+    model.add(tf.keras.layers.Conv2D(16, (3,3), padding='same', activation='relu'))
+    model.add(tf.keras.layers.BatchNormalization(momentum=momentum))
+    
+    model.add(tf.keras.layers.MaxPool2D())
+    
+    model.add(tf.keras.layers.Conv2D(32, (3,3), padding='same', activation='relu'))
+    model.add(tf.keras.layers.Conv2D(32, (3,3), padding='same', activation='relu'))
+    model.add(tf.keras.layers.BatchNormalization(momentum=momentum))
+    
+    # flatten...
+    model.add(tf.keras.layers.GlobalMaxPool2D())
     return model
 
-
-def create_model_lstm():
-
-    shape=(5, 90, 160, 3)
-    
+def action_model(shape=(w, 135, 240, 3), nbout=1):
     # Create our convnet with (112, 112, 3) input shape
-    convnet = create_model()
+    convnet = build_convnet(shape[1:])
     
     # then create our final model
-    lstm_model = tf.keras.Sequential()
+    model = tf.keras.Sequential()
     # add the convnet with (5, 112, 112, 3) shape
-    lstm_model.add(tf.keras.layers.TimeDistributed(convnet, input_shape=shape))
+    model.add(tf.keras.layers.TimeDistributed(convnet, input_shape=shape))
     # here, you can also use GRU or LSTM
-    lstm_model.add(tf.keras.layers.LSTM(64))
+    model.add(tf.keras.layers.GRU(64))
     # and finally, we make a decision network
-    lstm_model.add(tf.keras.layers.Dense(1024, activation='relu'))
-    lstm_model.add(tf.keras.layers.Dropout(.5))
-    lstm_model.add(tf.keras.layers.Dense(512, activation='relu'))
-    lstm_model.add(tf.keras.layers.Dropout(.5))
-    lstm_model.add(tf.keras.layers.Dense(128, activation='relu'))
-    lstm_model.add(tf.keras.layers.Dropout(.5))
-    lstm_model.add(tf.keras.layers.Dense(64, activation='relu'))
-    lstm_model.add(tf.keras.layers.Dense(1, activation='sigmoid'))
-
-    lstm_model.compile(
-        loss=tf.keras.losses.BinaryCrossentropy(),
-        optimizer=tf.keras.optimizers.Adam(lr=0.001/5),
-        metrics=[tf.keras.metrics.RecallAtPrecision(precision=0.9, name='recallAtPrecision'), 
-        tf.keras.metrics.BinaryAccuracy(threshold=0.5, name='binaryAccuracy')])
-
-    return lstm_model
-
-model = create_model_lstm()
-model.summary()
-
+    model.add(tf.keras.layers.Dense(32, activation='relu'))
+    model.add(tf.keras.layers.Dropout(.5))
+    model.add(tf.keras.layers.Dense(8, activation='relu'))
+    model.add(tf.keras.layers.Dense(nbout, activation='sigmoid'))
+    return model
 
 #%%
+
+model = action_model()
+model.summary()
+
+model.compile(
+        loss=tf.keras.losses.BinaryCrossentropy(),
+        optimizer=tf.keras.optimizers.Adam(),
+        metrics=[tf.keras.metrics.RecallAtPrecision(precision=0.9, name='recallAtPrecision'), 
+        tf.keras.metrics.BinaryAccuracy(threshold=0.6, name='binaryAccuracy')])
 
 cp_callback = tf.keras.callbacks.ModelCheckpoint(filepath=checkpoint_path,
                                                   save_weights_only=True, monitor='val_recallAtPrecision', verbose=1, 
                                                   save_best_only=True, mode='max')
 
-model.fit(x=dataset_train, validation_data=dataset_val, epochs=200, 
+model.fit(x=dataset_train, validation_data=dataset_val, epochs=50, 
                                 verbose=1,callbacks = [cp_callback], class_weight = {0: 1 , 1:1.92})
 
-#%%
+# %%
+
 print("Evaluate on test data")
 results = model.evaluate(dataset_test)
 print("test loss, test acc:", results)
@@ -286,10 +251,8 @@ print("train loss, trai acc:", results)
 model.load_weights(checkpoint_path)
 print("Evaluate on test data")
 results = model.evaluate(dataset_test)
-print("test loss, rap, rac:", results)
+print("test loss, test acc:", results)
 
 print("Evaluate on train data")
 results = model.evaluate(dataset_train)
-print("train loss, rap, rac:", results)
-
-
+print("train loss, trai acc:", results)
